requirements.txt
certifi==2020.12.5
Django==2.2.9
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==3.10.0
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
wincertstore==0.2
zipp==3.4.1
